[#324][BACK][TEST] - 도서 관련 서비스 테스트 코드 작성 (#425)
* feat(book) : 도서 작가 service 테스트 코드 작성

# Conflicts:
#	backend-server/src/test/java/shop/bluebooktle/backend/book/service/BookAuthorServiceTest.java

* feat(book) : 도서 이미지 service 테스트 코드 작성

* test(book): 도서 카테고리 서비스 테스트 코드 작성 (60%)

* test(user): 휴면 사용자 레디스 인증 코드 서비스 테스트 코드 작성 (100%)

* test(book): 로직 수정으로 인한 도서 관련 서비스 테스트 코드 주석 처리

* test(user): 사용자 서비스 테스트 코드 작성 (100%)

* test(user): AddressControllerTest, AddressControllerUnitTest 추가

* test(book): 로직 변경으로 인한 AuthorServiceTest 수정

* feat/test(book): BookAuthor 관련 서비스 메서드 정리 및 테스트 코드 수정

* test(book): AladinBook관련 서비스 메서드 수정 및 테스트 코드 추가

* cmt

* test(book): 도서 이미지 관련 서비스 테스트 코드 작성

---------

+++ backend-server/src/main/java/shop/bluebooktle/backend/book/service/impl/BookImgServiceImpl.java
@@ -1,7 +1,6 @@
 package shop.bluebooktle.backend.book.service.impl;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,19 +13,10 @@ import shop.bluebooktle.backend.book.repository.BookImgRepository;
 import shop.bluebooktle.backend.book.repository.BookRepository;
 import shop.bluebooktle.backend.book.repository.ImgRepository;
 import shop.bluebooktle.backend.book.service.BookImgService;
-import shop.bluebooktle.backend.book.service.ImgService;
-import shop.bluebooktle.common.dto.book.request.BookImgRegisterRequest;
-import shop.bluebooktle.common.dto.book.request.img.ImgRegisterRequest;
-import shop.bluebooktle.common.dto.book.response.BookImgResponse;
 import shop.bluebooktle.common.dto.book.response.img.ImgResponse;
 import shop.bluebooktle.common.exception.book.BookIdNullException;
-import shop.bluebooktle.common.exception.book.BookImgAlreadyExistsException;
 import shop.bluebooktle.common.exception.book.BookImgNotFoundException;
 import shop.bluebooktle.common.exception.book.BookNotFoundException;
-import shop.bluebooktle.common.exception.book.ImgIdNullException;
-import shop.bluebooktle.common.exception.book.ImgNotFoundException;
-
-// 수정 필요
 
 @Service
 @Transactional
@@ -36,41 +26,14 @@ public class BookImgServiceImpl implements BookImgService {
 	private final BookRepository bookRepository;
 	private final ImgRepository imgRepository;
 	private final BookImgRepository bookImgRepository;
-	private final ImgService imgService;
-
-	@Override
-	public void registerBookImg(Long bookId, BookImgRegisterRequest bookImgRegisterRequest) {
-		if (bookId == null) {
-			throw new BookIdNullException();
-		}
-		imgService.registerImg(ImgRegisterRequest
-			.builder()
-			.imgUrl(bookImgRegisterRequest.getImgUrl())
-			.build());
-
-		Long imgId = bookImgRegisterRequest.getImgId();
-		if (imgId == null) {
-			throw new ImgIdNullException();
-		}
-
-		Img img = imgRepository.findById(bookImgRegisterRequest.getImgId())
-			.orElseThrow(() -> new ImgNotFoundException());
-		Book book = bookRepository.findById(bookId).orElseThrow(() -> new BookNotFoundException());
-
-		if (bookImgRepository.existsByBookAndImg(book, img)) {
-			throw new BookImgAlreadyExistsException(bookId, imgId);
-		}
-		BookImg bookImg = BookImg.builder()
-			.book(book)
-			.img(img)
-			.isThumbnail(bookImgRegisterRequest.isThumbnail())
-			.build();
-
-		bookImgRepository.save(bookImg);
-	}
 
 	@Override
 	public void registerBookImg(Long bookId, String imageUrl) {
+
+		if (bookId == null || imageUrl == null || imageUrl.trim().isEmpty()) {
+			throw new IllegalArgumentException();
+		}
+
 		Book book = bookRepository.findById(bookId).orElseThrow(BookNotFoundException::new);
 		Img img = imgRepository.save(Img.builder().imgUrl(imageUrl).build());
 
@@ -101,48 +64,16 @@ public class BookImgServiceImpl implements BookImgService {
 			.build();
 	}
 
-	@Transactional(readOnly = true)
-	@Override
-	public Optional<BookImgResponse> getThumbnailByBookId(Long bookId) {
-		if (bookId == null) {
-			throw new BookIdNullException();
-		}
-		/*return bookImgRepository.findByBookId(bookId).stream()
-			.filter(BookImg::isThumbnail)
-			.findFirst()
-			.map(bookImg -> BookImgResponse.builder()
-				.imgResponse(ImgResponse.builder()
-					.id(bookImg.getImg().getId())
-					.imgUrl(bookImg.getImg().getImgUrl())
-					.createdAt(bookImg.getImg().getCreatedAt())
-					.build()
-				)
-				.bookInfoResponse(new BookInfoResponse(bookImg.getBook().getId()))
-				.isThumbnail(true)
-				.build()
-			);*/
-		return null;
-	}
-
-	@Override
-	public void deleteBookImg(Long bookId, Long imgId) {
-		if (bookId == null) {
-			throw new BookIdNullException();
-		}
-		if (imgId == null) {
-			throw new ImgIdNullException();
-		}
-
-		BookImg bookImg = bookImgRepository.findByBookIdAndImgId(bookId, imgId)
-			.orElseThrow(() -> new BookImgNotFoundException(bookId, imgId));
-
-		bookImgRepository.delete(bookImg);
-	}
-
 	@Override
 	public void updateBookImg(Long bookId, String imageUrl) {
 		// 도서에 기존 이미지 삭제
+
 		List<BookImg> bookImgList = bookImgRepository.findByBookId(bookId);
+
+		if (bookImgList.isEmpty()) {
+			throw new BookImgNotFoundException();
+		}
+
 		bookImgRepository.deleteAll(bookImgList);
 		// 새로운 이미지 등록
 		registerBookImg(bookId, imageUrl);
